+++ lib/modules/stopwatch_isolate/src/stopwatch_isolate_model.dart
@@ -2,5 +2,5 @@ import '../../custom_isolate/custom_isolate.dart';
 
 class StopwatchIsolateModel extends IsolateModel {
   final Stopwatch stopwatch = Stopwatch();
-  StopwatchIsolateModel();
+  StopwatchIsolateModel({required super.isolateName});
 }