+++ CHANGELOG.md
@@ -2,3 +2,8 @@
 
 * Add custom isolate module
 * Add stopwatch isolate module
+## 0.0.2
+
+* Add field `isolateName` to Isolate Model
+* Change return type of stopwatch from `int` to `Duration`
+* Add github workflows to anaylze the code